+++ app.js
@@ -55,9 +55,6 @@ var myprofileRouter = require('./routes/myprofile');
 var myprofileEditRouter = require('./routes/myprofileEdit');
 var myprofileUpdateRouter = require('./routes/myprofileUpdate');
 var myprofile_imgRouter = require('./routes/myprofile_img');
-var myprofile_img2Router = require('./routes/myprofile_img2');
-var myprofile_img3Router = require('./routes/myprofile_img3');
-var myprofile_img4Router = require('./routes/myprofile_img4');
 var registerRouter = require('./routes/register');
 var loginRouter = require('./routes/login');
 var logoutRouter = require('./routes/logout');
@@ -96,9 +93,6 @@ app.use('/myprofile', myprofileRouter);
 app.use('/myprofileEdit', myprofileEditRouter);
 app.use('/myprofileUpdate', myprofileUpdateRouter);
 app.use('/myprofile_img', myprofile_imgRouter);
-app.use('/myprofile_img2', myprofile_img2Router);
-app.use('/myprofile_img3', myprofile_img3Router);
-app.use('/myprofile_img4', myprofile_img4Router);
 app.use('/register', registerRouter);
 app.use('/login', loginRouter);
 app.use('/logout', logoutRouter);
