+++ routes/diaryDelete.js
@@ -1,12 +1,8 @@
 var express = require('express');
 var router = express.Router();
-const connection = require('../mysqlConnection');//db接続読み取り
+let diary_c = require('../controllers/diary')//コントローラーフォルダのhogeファイルを読み取る
 
 //削除のルーティング
-router.get('/:id',(req,res)=>{
-    connection.query('DELETE FROM diary WHERE id = ?',[req.params.id],(error,results)=>{
-      res.redirect('/drege')
-    })
-  })
+router.get('/:id',diary_c.diaryDelete)
 
   module.exports = router;
