+++ routes/drege.js
@@ -1,18 +1,9 @@
 var express = require('express');
 var router = express.Router();
-const connection = require('../mysqlConnection');//db接続読み取り
+
+let diary_c = require('../controllers/diary')//コントローラーフォルダの指定ファイルを読み取る
 
 
-
-router.get('/', (req, res) => {
-  // var body = req.body;
-  // console.log(body);
-  var userId = req.session.user_id? req.session.user_id: 0;
-  var sql = 'SELECT D.id, D.user_id, D.title, D.image, D.tag, D.text,ifnull(U.name, \'名無し\') AS name, DATE_FORMAT(D.created_at, \'%Y年%m月%d日 %k時%i分%s秒\') AS created_at FROM diary D LEFT OUTER JOIN user U ON D.user_id = U.user_id WHERE D.user_id = ' + userId + '  ORDER BY D.created_at DESC'; // 変更
-  connection.query(sql,(error,results)=>{
-      res.render('drege.ejs',{items:results})
-      console.log(results)
-    })
-})	;
+router.get('/',diary_c.drege)	;
 
 module.exports = router;
