+++ routes/flowerDelete.js
@@ -1,12 +1,9 @@
 var express = require('express');
 var router = express.Router();
-const connection = require('../mysqlConnection');//db接続読み取り
+
+let flower_c = require('../controllers/flower')//コントローラーフォルダの指定ファイルを読み取る
 
 //削除のルーティング
-router.get('/:id',(req,res)=>{
-    connection.query('DELETE FROM flower WHERE id = ?',[req.params.id],(error,results)=>{
-      res.redirect('/frege')
-    })
-  })
+router.get('/:id',flower_c.flowerDelete)
 
   module.exports = router;
