+++ routes/mypage.js
@@ -1,13 +1,7 @@
 var express = require('express');
 var router = express.Router();
-const connection = require('../mysqlConnection');//db接続読み取り
+let shop_c = require('../controllers/shop')//コントローラーフォルダの指定ファイルを読み取る
 
-router.get('/', (req, res) => {
- 
-    var sql = 'SELECT * FROM shop'; // 変更
-    connection.query(sql,(error,results)=>{
-        res.render('mypage.ejs',{items:results})
-        // console.log(results)
-      })
-  })	;
+router.get('/', shop_c.mypage);
+
 module.exports = router;
