+++ routes/top.js
@@ -1,15 +1,8 @@
 var express = require('express');
 var router = express.Router();
-const connection = require('../mysqlConnection');//db接続読み取り
 
-router.get('/',(req, res)=>{
-connection.query('SELECT * FROM user',(error,results)=>{
-  // res.renderで指定ファイルの画面表示させる
-  res.render('top.ejs');
-})
+let main_c = require('../controllers/main')//コントローラーフォルダの指定ファイルを読み取る
 
-
-  
-});
+router.get('/',main_c.top);
 
 module.exports = router;
